requirements_new.txt
# Core Django - Using Django 5.1 for better compatibility
Django==5.1.5
djangorestframework==3.15.2
django-allauth==65.9.0
django-cors-headers==4.7.0
django-filter==24.3
django-extensions==3.2.3
django-debug-toolbar==4.4.6
django-storages==1.14.6

# Database
dj-database-url==3.0.0
psycopg2-binary==2.9.10

# Environment management
python-decouple==3.8

# Async/Channels
channels==4.1.0
channels-redis==4.2.1
asgiref==3.8.1

# Task queue
celery==5.4.0
redis==5.2.1

# Utilities
Pillow==11.2.1
requests==2.32.4
python-dateutil==2.9.0.post0

# Development
ipython==8.26.0

# AWS (opcional)
boto3==1.38.36
botocore==1.38.36

# Additional packages for advanced features
django-elasticsearch-dsl==8.0
elasticsearch==8.15.1
django-haystack==3.3.0
whoosh==2.7.4

# Full-text search and indexing
django-watson==1.6.3

# Monitoring and performance
django-silk==5.1.0

# Advanced permissions
django-guardian==2.4.0

# Rich text editing
django-ckeditor==6.7.1

# API documentation
drf-spectacular==0.28.0

# Rate limiting
django-ratelimit==4.1.0

# Background tasks - Compatible with Django 5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
